src/main/java/com/igor/hexagonal/application/core/usecase/InsertCustomerUseCase.java
package com.igor.hexagonal.application.core.usecase;

import com.igor.hexagonal.application.core.domain.Customer;
import com.igor.hexagonal.application.ports.in.InsertCustomerInputPort;
import com.igor.hexagonal.application.ports.out.FindAddressByZipCodeOutputPort;
import com.igor.hexagonal.application.ports.out.InsertCustomerOutputPort;

public class InsertCustomerUseCase implements InsertCustomerInputPort {

    private final FindAddressByZipCodeOutputPort findAddressByZipCodeOutputPort;
    private final InsertCustomerOutputPort insertCustomerOutputPort;

    public InsertCustomerUseCase(FindAddressByZipCodeOutputPort findAddressByZipCodeOutputPort, InsertCustomerOutputPort insertCustomerOutputPort) {
        this.findAddressByZipCodeOutputPort = findAddressByZipCodeOutputPort;
        this.insertCustomerOutputPort = insertCustomerOutputPort;
    }

    @Override
    public void insert(final Customer customer, final String zipCode) {
        var address = findAddressByZipCodeOutputPort.find(zipCode);
        customer.setAddress(address);

        insertCustomerOutputPort.insert(customer);
    }
}
